+++ continew-starter-extension/continew-starter-extension-crud/continew-starter-extension-crud-core/src/main/java/top/continew/starter/extension/crud/controller/AbstractBaseController.java
@@ -30,7 +30,6 @@ import top.continew.starter.extension.crud.enums.Api;
 import top.continew.starter.extension.crud.handler.CrudApiHandler;
 import top.continew.starter.extension.crud.model.query.PageQuery;
 import top.continew.starter.extension.crud.model.query.SortQuery;
-import top.continew.starter.extension.crud.model.req.BaseReq;
 import top.continew.starter.extension.crud.model.resp.BaseIdResp;
 import top.continew.starter.extension.crud.model.resp.BasePageResp;
 import top.continew.starter.extension.crud.service.BaseService;
@@ -49,7 +48,7 @@ import java.util.List;
  * @author Charles7c
  * @since 1.0.0
  */
-public abstract class AbstractBaseController<S extends BaseService<L, D, Q, C>, L, D, Q, C extends BaseReq> implements CrudApiHandler {
+public abstract class AbstractBaseController<S extends BaseService<L, D, Q, C>, L, D, Q, C> implements CrudApiHandler {
 
     @Autowired
     protected S baseService;
@@ -65,7 +64,7 @@ public abstract class AbstractBaseController<S extends BaseService<L, D, Q, C>, 
     @Operation(summary = "分页查询列表", description = "分页查询列表")
     @ResponseBody
     @GetMapping
-    public BasePageResp<L> page(Q query, @Validated PageQuery pageQuery) {
+    public BasePageResp<L> page(@Validated Q query, @Validated PageQuery pageQuery) {
         return baseService.page(query, pageQuery);
     }
 
@@ -80,7 +79,7 @@ public abstract class AbstractBaseController<S extends BaseService<L, D, Q, C>, 
     @Operation(summary = "查询列表", description = "查询列表")
     @ResponseBody
     @GetMapping("/list")
-    public List<L> list(Q query, SortQuery sortQuery) {
+    public List<L> list(@Validated Q query, @Validated SortQuery sortQuery) {
         return baseService.list(query, sortQuery);
     }
 
@@ -95,7 +94,7 @@ public abstract class AbstractBaseController<S extends BaseService<L, D, Q, C>, 
     @Operation(summary = "查询树列表", description = "查询树列表")
     @ResponseBody
     @GetMapping("/tree")
-    public List<Tree<Long>> tree(Q query, SortQuery sortQuery) {
+    public List<Tree<Long>> tree(@Validated Q query, @Validated SortQuery sortQuery) {
         return baseService.tree(query, sortQuery, false);
     }
 
@@ -168,7 +167,7 @@ public abstract class AbstractBaseController<S extends BaseService<L, D, Q, C>, 
     @ExcludeFromGracefulResponse
     @Operation(summary = "导出数据", description = "导出数据")
     @GetMapping("/export")
-    public void export(Q query, SortQuery sortQuery, HttpServletResponse response) {
+    public void export(@Validated Q query, @Validated SortQuery sortQuery, HttpServletResponse response) {
         baseService.export(query, sortQuery, response);
     }
 }